backend/pyproject.toml
[tool.poetry]
name = "uncons-backend"
version = "0.1.0"
description = "Voice-first AI tutor for Anki flashcards - Backend"
authors = ["Lisardo Iniesta"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-dotenv = "^1.0.0"
httpx = "^0.28.0"
websockets = "^14.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"

# Voice pipeline
deepgram-sdk = "^3.8.0"
livekit = "^1.0.23"
livekit-agents = "^1.3.10"
livekit-plugins-deepgram = "^1.3.10"
livekit-plugins-silero = "^1.3.10"
livekit-plugins-cartesia = "^1.3.10"
livekit-plugins-openai = "^1.3.10"

# AI/LLM
google-generativeai = "^0.8.0"
openai = "^1.60.0"
langgraph = "^0.2.0"
langgraph-checkpoint-sqlite = "^2.0.0"
av = "^16.0.1"
tenacity = "^9.1.2"
livekit-plugins-google = "^1.3.10"

# Message ID generation
python-ulid = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
pytest-dotenv = "^0.5.2"
ruff = "^0.8.0"
black = "^24.10.0"
mypy = "^1.13.0"
pre-commit = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "SIM102", # Nested if statements (sometimes clearer separate)
    "SIM116", # Dict instead of if-elif (sometimes clearer explicit)
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
env_files = ["../.env"]
markers = [
    "integration: marks tests requiring real services (deselect with '-m not integration')",
    "integration_fast: marks fast integration tests (no Docker/external services required)",
    "e2e: marks end-to-end tests",
    "llm_judge: marks tests using real LLM for evaluation (expensive, use sparingly)",
    "slow: marks slow tests requiring API calls (deselect with '-m not slow')",
]
